erstatt enkelt interface med implementasjon
