+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/Config.kt
@@ -2,7 +2,7 @@ package no.nav.mulighetsrommet.api
 
 import io.ktor.client.engine.*
 import io.ktor.client.engine.cio.*
-import no.nav.mulighetsrommet.api.clients.brreg.BrregClientImpl
+import no.nav.mulighetsrommet.api.clients.brreg.BrregClient
 import no.nav.mulighetsrommet.api.clients.sanity.SanityClient
 import no.nav.mulighetsrommet.api.domain.dbo.NavAnsattRolle
 import no.nav.mulighetsrommet.api.tasks.*
@@ -36,7 +36,7 @@ data class AppConfig(
     val tasks: TaskConfig,
     val norg2: Norg2Config,
     val slack: SlackConfig,
-    val brreg: BrregClientImpl.Config,
+    val brreg: BrregClient.Config,
     val unleash: UnleashService.Config,
     val axsys: ServiceClientConfig,
     val engine: HttpClientEngine = CIO.create(),
